Drop Base deploy script and tidy Foundry mocks and tests

+++ packages/foundry/test/invariant/Invariant.t.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.28;
 
-import {Test} from "forge-std/Test.sol";
-import {TradeLayer} from "../../contracts/TradeLayer.sol";
-import {MockUSDC} from "../../contracts/mocks/MockUSDC.sol";
-import {Handler} from "./Handler.t.sol";
+import { Test } from "forge-std/Test.sol";
+import { TradeLayer } from "../../contracts/TradeLayer.sol";
+import { MockUSDC } from "../../contracts/mocks/MockUSDC.sol";
+import { Handler } from "./Handler.t.sol";
 
 contract TradeLayerInvariantTest is Test {
     TradeLayer public tradeLayer;
     MockUSDC public usdc;
     Handler public handler;
-    
+
     function setUp() public {
         // Deploy contracts
         usdc = new MockUSDC(); // this contract gets minted 10 million mock usdc
@@ -33,40 +33,29 @@ contract TradeLayerInvariantTest is Test {
         selectors[2] = Handler.redeemStock.selector;
         selectors[3] = Handler.fulfillRedeemRequest.selector;
 
-        targetSelector(FuzzSelector({
-            addr: address(handler),
-            selectors: selectors
-        }));
+        targetSelector(FuzzSelector({ addr: address(handler), selectors: selectors }));
     }
-    
+
     /* ---------- INVARIANT 1: USDC CONSERVATION ---------- */
-    
+
     /// @notice Contract must hold enough USDC to cover all pending buy requests
     function invariant_usdcConservation() public view {
         uint256 contractBalance = usdc.balanceOf(address(tradeLayer));
         uint256 totalPendingUsdc = handler.getTotalPendingUsdc();
-        
-        assertGe(
-            contractBalance, 
-            totalPendingUsdc, 
-            "INVARIANT VIOLATED: Contract USDC < Pending USDC"
-        );
+
+        assertGe(contractBalance, totalPendingUsdc, "INVARIANT VIOLATED: Contract USDC < Pending USDC");
     }
-    
+
     /* ---------- INVARIANT 2: TOKEN SUPPLY = TOTAL HOLDINGS ---------- */
-    
+
     /// @notice Total supply of DSTOCK must equal sum of all user holdings
     function invariant_tokenSupplyEqualsHoldings() public view {
         uint256 totalSupply = tradeLayer.totalSupply();
         uint256 sumOfHoldings = handler.getSumOfAllHoldings();
-        
-        assertEq(
-            totalSupply, 
-            sumOfHoldings, 
-            "INVARIANT VIOLATED: Total Supply != Sum of Holdings"
-        );
+
+        assertEq(totalSupply, sumOfHoldings, "INVARIANT VIOLATED: Total Supply != Sum of Holdings");
     }
-    
+
     /* ---------- INVARIANT 3: SETTLEMENT IS BOUND TO INTENT ---------- */
 
     /// @notice The contract's recorded per-user, per-stock holdings must equal the
@@ -88,45 +77,42 @@ contract TradeLayerInvariantTest is Test {
             }
         }
     }
-    
+
     /* ---------- INVARIANT 4: USER BALANCE >= USER HOLDINGS ---------- */
-    
+
     /// @notice Each user's token balance must be >= their total stock holdings
     function invariant_userBalanceCoversHoldings() public view {
         address[] memory actors = _getActors();
         string[] memory stocks = _getStocks();
-        
-        for (uint i = 0; i < actors.length; i++) {
+
+        for (uint256 i = 0; i < actors.length; i++) {
             uint256 balance = tradeLayer.balanceOf(actors[i]);
             uint256 totalHoldings = 0;
-            
-            for (uint j = 0; j < stocks.length; j++) {
+
+            for (uint256 j = 0; j < stocks.length; j++) {
                 totalHoldings += tradeLayer.totalHoldings(actors[i], stocks[j]);
             }
-            
+
             assertGe(
-                balance, 
-                totalHoldings, 
-                string(abi.encodePacked(
-                    "INVARIANT VIOLATED: User balance < holdings for ",
-                    vm.toString(actors[i])
-                ))
+                balance,
+                totalHoldings,
+                string(abi.encodePacked("INVARIANT VIOLATED: User balance < holdings for ", vm.toString(actors[i])))
             );
         }
     }
-    
+
     /* ---------- INVARIANT 5: TRANSFERS ALWAYS REVERT ---------- */
-    
+
     /// @notice Transfer and transferFrom should always revert
     function invariant_transfersDisabled() public {
         address[] memory actors = _getActors();
-        
-        for (uint i = 0; i < actors.length; i++) {
+
+        for (uint256 i = 0; i < actors.length; i++) {
             uint256 balance = tradeLayer.balanceOf(actors[i]);
-            
+
             if (balance > 0 && actors.length > 1) {
                 address recipient = actors[(i + 1) % actors.length];
-                
+
                 vm.prank(actors[i]);
                 try tradeLayer.transfer(recipient, 1) {
                     fail("INVARIANT VIOLATED: Transfer should revert");
@@ -140,7 +126,7 @@ contract TradeLayerInvariantTest is Test {
             }
         }
     }
-    
+
     /* ---------- INVARIANT 6: ACCOUNTING CONSISTENCY ---------- */
 
     /// @notice Ghost variable tracking should match actual state
@@ -149,9 +135,7 @@ contract TradeLayerInvariantTest is Test {
         uint256 netSupply = handler.ghost_totalTokensMinted() - handler.ghost_totalTokensBurned();
 
         assertEq(
-            netSupply,
-            tradeLayer.totalSupply(),
-            "INVARIANT VIOLATED: Ghost tracking inconsistent with actual supply"
+            netSupply, tradeLayer.totalSupply(), "INVARIANT VIOLATED: Ghost tracking inconsistent with actual supply"
         );
     }
 
@@ -184,7 +168,6 @@ contract TradeLayerInvariantTest is Test {
         );
     }
 
-
     /* ---------- HELPERS ---------- */
 
     // Delegate to the handler so the actor/stock sets cannot drift apart.
@@ -195,9 +178,9 @@ contract TradeLayerInvariantTest is Test {
     function _getStocks() internal view returns (string[] memory) {
         return handler.getStocks();
     }
-    
+
     /* ---------- AFTER INVARIANT ---------- */
-    
+
     function invariant_callSummary() public view {
         handler.callSummary();
     }
